feat: enhance customer service and error handling with new response structures and transaction management

+++ src/main/java/com/ojo/cyrus/models/responses/SubAccountBalanceResponse.java
@@ -1,12 +1,12 @@
 package com.ojo.cyrus.models.responses;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.Instant;
 
 public record SubAccountBalanceResponse(
         String accountId,
         String accountType,
-        BigDecimal balance,
+        BigInteger balance,   // integer kobo (minor units); convert to naira at the display edge
         String currency,
         Instant asOf
 ) {}
